Clarify type identity and conversion
I've noticed a lot of people get confused with how sort.Sort() Works.
Defining the interface and how the type satisfies it is only half of the
equation. It's also helpful to describe *why* the new type is created in
the first place, which is to serve both as a place to hang the methods
necessary to implement sort.Interface as well as provide meaningful
syntax in the sort statement itself. By convention, they start with `by`
to make the statement easier to read.